fixed un-needed binding redundancies
